+++ src/main/java/com/nemonotfound/nemosfarming/NemosFarming.java
@@ -1,6 +1,10 @@
 package com.nemonotfound.nemosfarming;
 
+import com.nemonotfound.nemosfarming.block.ModBlocks;
 import com.nemonotfound.nemosfarming.enchantment.ModEnchantments;
+import com.nemonotfound.nemosfarming.item.ModItemGroups;
+import com.nemonotfound.nemosfarming.item.ModItems;
+import com.nemonotfound.nemosfarming.loot.ModLootTables;
 import net.fabricmc.api.ModInitializer;
 
 import org.slf4j.Logger;
@@ -16,5 +20,9 @@ public class NemosFarming implements ModInitializer {
 		log.info("Thank you for using Nemo's Farming!");
 
 		ModEnchantments.registerEnchantments();
+		ModBlocks.registerBlocks();
+		ModItems.registerItems();
+		ModItemGroups.registerItemGroups();
+		ModLootTables.registerLootTables();
 	}
 }