+++ src/main/java/com/nemonotfound/nemosfarming/NemosFarmingClient.java
@@ -1,0 +1,16 @@
+package com.nemonotfound.nemosfarming;
+
+import com.nemonotfound.nemosfarming.block.ModBlocks;
+import net.fabricmc.api.ClientModInitializer;
+import net.fabricmc.fabric.api.blockrenderlayer.v1.BlockRenderLayerMap;
+import net.minecraft.client.render.RenderLayer;
+
+public class NemosFarmingClient implements ClientModInitializer {
+
+	@Override
+	public void onInitializeClient() {
+		BlockRenderLayerMap.INSTANCE.putBlock(ModBlocks.LETTUCE, RenderLayer.getCutout());
+		BlockRenderLayerMap.INSTANCE.putBlock(ModBlocks.TOMATO, RenderLayer.getCutout());
+		BlockRenderLayerMap.INSTANCE.putBlock(ModBlocks.CUCUMBER, RenderLayer.getCutout());
+	}
+}