+++ src/main/java/com/nemonotfound/nemosfarming/NemosFarmingDataGenerator.java
@@ -1,7 +1,6 @@
 package com.nemonotfound.nemosfarming;
 
-import com.nemonotfound.nemosfarming.datagen.EnchantmentGenerator;
-import com.nemonotfound.nemosfarming.datagen.EnchantmentTagGenerator;
+import com.nemonotfound.nemosfarming.datagen.*;
 import com.nemonotfound.nemosfarming.datagen.langdatagen.EnglishLanguageGenerator;
 import com.nemonotfound.nemosfarming.datagen.langdatagen.GermanLanguageGenerator;
 import net.fabricmc.fabric.api.datagen.v1.DataGeneratorEntrypoint;
@@ -17,5 +16,9 @@ public class NemosFarmingDataGenerator implements DataGeneratorEntrypoint {
 		pack.addProvider(GermanLanguageGenerator::new);
 		pack.addProvider(EnchantmentGenerator::new);
 		pack.addProvider(EnchantmentTagGenerator::new);
+		pack.addProvider(ModelProvider::new);
+		pack.addProvider(LootTableGnerator::new);
+		pack.addProvider(RecipeGenerator::new);
+		pack.addProvider(ItemTagProvider::new);
 	}
 }