+++ src/main/java/com/nemonotfound/nemosfarming/block/CucumberCropBlock.java
@@ -1,0 +1,39 @@
+package com.nemonotfound.nemosfarming.block;
+
+import com.nemonotfound.nemosfarming.item.ModItems;
+import net.minecraft.block.Block;
+import net.minecraft.block.BlockState;
+import net.minecraft.block.CropBlock;
+import net.minecraft.item.ItemConvertible;
+import net.minecraft.state.StateManager;
+import net.minecraft.state.property.IntProperty;
+
+public class CucumberCropBlock extends CropBlock {
+
+    public static final int MAX_AGE = 3;
+    public static final IntProperty AGE = IntProperty.of("age", 0, 3);
+
+    public CucumberCropBlock(Settings settings) {
+        super(settings);
+    }
+
+    @Override
+    protected ItemConvertible getSeedsItem() {
+        return ModItems.CUCUMBER_SEEDS;
+    }
+
+    @Override
+    public IntProperty getAgeProperty() {
+        return AGE;
+    }
+
+    @Override
+    public int getMaxAge() {
+        return MAX_AGE;
+    }
+
+    @Override
+    protected void appendProperties(StateManager.Builder<Block, BlockState> builder) {
+        builder.add(AGE);
+    }
+}