+++ Changelog.md
@@ -1,9 +1,37 @@
-# Changelog Snapshot-24w32a
+# Changelog v1.0
 
 ## Additions
-- Added enchantment Replanting for hoes
-  - Replant crops
+- Added crops
+  - Tomato
+    - Seeds can be bought from the wandering trader
+    - Can be sold to farmer
+    - Can be found in chests
+  - Lettuce
+    - Seeds can be bought from the wandering trader
+    - Can be sold to farmer
+    - Can be found in chests
+  - Cucumber
+    - Seeds can be bought from the wandering trader
+    - Can be sold to farmer
+    - Can be found in chests
+- Added food
+  - Mixes Salad
+  - Mixed Salad with Carrots
+  - Mixed Salad with Beetroot
+- Added item group Nemo's Farming
+
+## Changes
+ - Reaper's max. level is now 3
+   - level 1: 3x3x3
+   - level 2: 5x5x5
+   - level 3: 7x7x7
+
+## Breaking Changes
+- Fixed identifier of Farmer's Knowledge
+  - This causes the enchantment to vanish
+
+For recipes and more info check the [wiki](https://nemonotfound.com/minecraft-mods/nemos-farming/wiki)
 
 ### Downloads
 [CurseForge](https://curseforge.com/minecraft/mc-mods/nemos-farming) <br>
-[Modrinth](https://modrinth.com/mod/nemos-nemos-farming)
+[Modrinth](https://modrinth.com/mod/nemos-farming)